conda_requirements.txt
ephem
pillow
pyiem
pytest-cov
pytest-runner
pytest
requests
